ref: Move response functions to their own module
